[IMP] sale_timesheet: add tour js to check the process in the UI
This tour js is used to check if the process is not broken. To do this,
we create an unit test to launch this tour js.

Before launching this test we create a prepaid service product in the
unit test.

Steps of this tour js:

1) Go to the Sales app,
2) Create a Sales Order with the prepaid service product and we order 10
hours for this product.
3) Go to the Project App,
4) Create a billable project to timesheet for the customer who orders
these hours,
5) Create task in this project and add a timesheet into it,
6) Show in this task, we can see the Sales Order Item linked to the
created timesheet, because this field is hidden by default.
7) Go to the Settings > Projects,
8) Create a new project and show the three different
configurations to invoice services in a project. That is, 'task rate',
'project rate' and 'employee rate'.

task-2439304