[REV] mail: fix a traceback when sending a message in a "email channel"
Bug
===
Since 1b78adbaae472d71df917ac2c40ababc3bcea0ad , if multiple users of
different companies are in the same channel, and if the type of the
channel is "email", a traceback is raised when sending a message.

Technical
=========
"with_user" drop the SU flag and is not enough to bypass the multi
company ACLs. Moreover, we do not want to send a "channel_seen" bus
notifications and we can mark the message as seen on the JS side
directly.

Changes
=======
So now, the traceback is not raised anymore in that case, and the
message is marked as seen if the receiver has the web client open.

But, if the receiver has not the web client open, when he will login,
the message will not be marked as seen. This can be improved but at the
moment we just want to revert the fix that cause the traceback and fix
the issue in most cases.

Task 2427790

closes odoo/odoo#71320

X-original-commit: 65fa6fe81ffd16ee7f0775056000b15ba10a39ed
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>