[FIX] web_editor: properly hide the custom colors section if needed
The 'custom' key was not working anymore as a key to exclude the
ability to choose a custom color from a color palette widget. It was
thus appearing for choosing the transparent color for over-the-content
headers... which is not possible to save right now.

Part of https://github.com/odoo/odoo/pull/65139

closes odoo/odoo#65154

X-original-commit: 8f9eb57df5c6a253ee0702c0e1fb2016cab7eb1e
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>