[IMP] account: add bank account step in invoices setup bar for community
This way, the bank account can be easily and immediately configured so that QR-codes can be generated on invoices without any additional setup (for Switzerland and SEPA).

This configuration step is hidden if full accounting features are enabled, since it will then appear in the "main" setup bar, on the dashboard.

Task 2412391